Bump numpy from 1.23.0 to 2.0.2

+++ script.py
@@ -4,7 +4,7 @@ def create_and_sum_array():
     """
     np.int32 を使用して NumPy 配列を作成し、その要素を足し合わせる関数
     """
-    arr = np.array([1, 2, 3], dtype=np.int)
+    arr = np.array([1, 2, 3], dtype=np.int32)
     total = np.sum(arr)
     return arr, total
   # 修正対象のファイル
